Добавил контекстное меню для видео в профиле, выделил ProfileVideoViewDelegate в отдельный файл (extension) и сделал несколько небольших изменений
