+++ XCE-FACTOR/View/CustomViews/ProfileVideoView.swift
@@ -7,11 +7,14 @@
 //
 
 import UIKit
+import AVKit
 import NVActivityIndicatorView
 
 protocol ProfileVideoViewDelegate: class {
     func playButtonPressed(at index: Int, video: Video)
     func optionsButtonPressed(at index: Int, video: Video)
+    func copyLinkButtonPressed(at index: Int, video: Video)
+    func shareButtonPreseed(at index: Int, video: Video)
 }
 
 class ProfileVideoView: UIView {
@@ -55,6 +58,12 @@ class ProfileVideoView: UIView {
         delegate?.optionsButtonPressed(at: index, video: video)
     }
     
+    //MARK:- Add Context Menu
+    func addContextMenu(delegate: UIContextMenuInteractionDelegate) {
+        let interaction = UIContextMenuInteraction(delegate: delegate)
+        self.addInteraction(interaction)
+    }
+    
 }
 
 private extension ProfileVideoView {
@@ -66,6 +75,8 @@ private extension ProfileVideoView {
             playButton.setImage(IconsManager.getIcon(.playSmall), for: .normal)
             optionsButton.setImage(IconsManager.getIcon(.optionDotsSmall), for: .normal)
         }
+        
+        self.addContextMenu(delegate: self)
     }
 
     private func xibSetup() {
@@ -74,4 +85,35 @@ private extension ProfileVideoView {
         contentView.frame = self.bounds
         contentView.autoresizingMask = [.flexibleHeight, .flexibleWidth]
     }
+}
+
+//MARK:- Context Menu Delegate
+///
+extension ProfileVideoView: UIContextMenuInteractionDelegate {
+    
+    //MARK:- Configure Menu
+    func contextMenuInteraction(_ interaction: UIContextMenuInteraction, configurationForMenuAtLocation location: CGPoint) -> UIContextMenuConfiguration? {
+
+        let config = UIContextMenuConfiguration(identifier: nil, previewProvider: nil) { (actions) -> UIMenu? in
+            
+            let copyLink = UIAction(title: "Скопировать ссылку", image: UIImage(systemName: "doc.on.doc")) { (action) in
+                self.delegate?.copyLinkButtonPressed(at: self.index, video: self.video)
+            }
+            
+            let shareVideo = UIAction(title: "Поделиться видео", image: UIImage(systemName: "square.and.arrow.up")) { (action) in
+                self.delegate?.shareButtonPreseed(at: self.index, video: self.video)
+            }
+            return UIMenu(title: "", children: [shareVideo, copyLink])
+            
+        }
+        return config
+    }
+    
+    //MARK:- Perform Preview Action
+    func contextMenuInteraction(_ interaction: UIContextMenuInteraction, willPerformPreviewActionForMenuWith configuration: UIContextMenuConfiguration, animator: UIContextMenuInteractionCommitAnimating) {
+        animator.addCompletion {
+            self.delegate?.playButtonPressed(at: self.index, video: self.video)
+        }
+    }
+    
 }
